Housekeeping * 'GameObject' now has a GetPosition function, so that classes that need that position don't have to get the 'Transform' and then access the position; * Some places where sf::vector2i were being used now use the wrapper vector2i; * One of the 'GameObject's constructor now calls another constructor that had the same definition, eliminating duplicate code; * 'Item' is now a 'Component'.

+++ StardewLike/src/InteractableTypes/NoInteractable.cpp
@@ -16,7 +16,7 @@ NoInteractable::NoInteractable(Vector2i aGridPosition)
 	PopulateInteractables();
 }
 
-NoInteractable::NoInteractable(float aGridPositionX, float aGridPositionY)
+NoInteractable::NoInteractable(int aGridPositionX, int aGridPositionY)
 	: NoInteractable(Vector2i(aGridPositionX, aGridPositionY))
 {}
